+++ Sources/KeyboardKit/_Keyboard/KeyboardController.swift
@@ -41,9 +41,6 @@ public protocol KeyboardController: AnyObject {
     /// Dismiss the keyboard.
     func dismissKeyboard()
 
-    /// Insert the provided autocomplete suggestion.
-    func insertAutocompleteSuggestion(_ suggestion: Autocomplete.Suggestion)
-
     /// Insert the provided text.
     func insertText(_ text: String)